versao04

+++ src/controller/clienteController.js
@@ -1,7 +1,6 @@
 const db = require('../db/db'); //módulo de conexão com o banco de dados
 const Joi = require('joi'); //Biblioteca de validação de dados
 const bcrypt = require('bcrypt'); //Para encriptação de senhas
-const { error } = require('console');
 
 //Validação com Joi
 const clienteSchema = Joi.object({
@@ -24,7 +23,7 @@ const clienteSchema = Joi.object({
 //Listar todos os clientes
 exports.listarClientes = async (req, res) => {
     try {
-        const[result] = await db.query('SELECT * FROM cliente');
+        const [result] = await db.query('SELECT * FROM cliente');
         res.json(result); //Aqui retornamos apenas os dados da consulta
     } catch (err) {
         console.error('Erro ao buscar clientes:', err);
@@ -38,7 +37,7 @@ exports.listarClienteCpf = async (req, res) => {
     try {
         const [result] = await db.query('SELECT * FROM cliente WHERE cpf = ?', [cpf]);
         if (result.length === 0) {
-            return res.status(404).json({ error: 'Cliente não encontrado '});
+            return res.status(404).json({ error: 'Cliente não encontrado ' });
         }
         res.json(result[0]);
     } catch (err) {
@@ -51,8 +50,13 @@ exports.listarClienteCpf = async (req, res) => {
 exports.deletarCliente = async (req, res) => {
     const { cpf } = req.params;
     try {
+        const [result] = await db.query('SELECT * FROM cliente WHERE cpf = ?', [cpf]);
+        if (result.length === 0) {
+            return res.status(404).json({ error: 'Cliente não encontrado' });
+        }
+
         await db.query('DELETE FROM cliente WHERE cpf = ?', [cpf]);
-        res.json({ message: 'Cliente deletado com sucesso '});
+        res.json({ message: 'Cliente deletado com sucesso ' });
     } catch (err) {
         console.error('Erro ao deletar cliente:', err);
         res.status(500).json({ error: 'Erro ao deletar cliente' });
@@ -68,29 +72,29 @@ exports.adicionarCliente = async (req, res) => {
     //Validação de dados
     const { error } = clienteSchema.validate({ cpf, nome, endereco, bairro, cidade, telefone, senha });
     if (error) {
-    return res.status(400),json({ error: error.details[0].message });
-  }
+        return res.status(400).json({ error: error.details[0].message });
+    }
 
-try{
-    //Criptografando a senha
-    const hash = await bcrypt.hash(senha, 10);
+    try {
+        //Criptografando a senha
+        const hash = await bcrypt.hash(senha, 10);
 
-    const novoCliente = { cpf, nome, endereco, bairro, cidade, telefone, senha: hash };
-    await db.query('INSERT INTO cliente SET ?' ,novoCliente);
+        const novoCliente = { cpf, nome, endereco, bairro, cidade, telefone, senha: hash };
+        await db.query('INSERT INTO cliente SET ?', novoCliente);
 
-    res.json({ message: 'Cliente adicionado com sucesso '});
-  } catch (err) {
-    console.log('Erro ao adicionar cliente:', err);
-    res.status(500).json({ error: 'Erro ao adicionar cliente' });
+        res.json({ message: 'Cliente adicionado com sucesso ' });
+    } catch (err) {
+        console.log('Erro ao adicionar cliente:', err);
+        res.status(500).json({ error: 'Erro ao adicionar cliente' });
 
-  }
+    }
 };
 
 
 //Atualizar um cliente
 exports.atualizarCliente = async (req, res) => {
     const { cpf } = req.params;
-    const {  nome, endereco, bairro, cidade, telefone, senha} = req.body;
+    const { nome, endereco, bairro, cidade, telefone, senha } = req.body;
 
     //Validação de dados
     const { error } = clienteSchema.validate({ cpf, nome, endereco, bairro, cidade, telefone, senha });
@@ -102,7 +106,7 @@ exports.atualizarCliente = async (req, res) => {
         //Criptografando a senha
         const hash = await bcrypt.hash(senha, 10);
 
-        const clienteAtualizado = {nome, endereco, bairro, cidade, telefone, senha: hash };
+        const clienteAtualizado = { nome, endereco, bairro, cidade, telefone, senha: hash };
         await db.query('UPDATE cliente SET ? WHERE cpf = ?', [clienteAtualizado, cpf]);
 
         res.json({ message: 'Cliente atualizado com sucesso' });
